python/requirements.txt
pillow
pytesseract
gtts
wand
